ice_history_shared: disallow 'x' in history frequency variables f_*'
In the current code, nothing prevents users from leaving 'x' along with
active frequencies in the individual namelist history frequency
variables, for example:

    f_aice = 'xmd'

This configuration does not work correctly, however. The corresponding
history fields are correctly defined in
ice_history_shared::define_hist_field, but since the calls to
ice_history_shared::accum_hist_field in ice_history::accum_hist are only
done after checking that the first element of each frequency variable is
not 'x', the corresponding variables in the history files are all zero.

Prevent that behaviour by actually disallowing 'x' in history frequency
variables if any other frequencies are active. To implement that, add a
check in the loop in define_hist_field, which loops through vhistfreq,
(corresponding to f_aice, etc. in ice_history). Since this subroutine
initializes 'id(:)' to zero and then writes a (non-zero) index in 'id'
for any active frequency, it suffices to check that all previous indices
are non-zero.