ice_history: use loop index directly for CMIP variables
In ice_history::accum_hist, there is a special treatment for some CMIP
variables where they are averaged only for time steps where ice is
present, and points where there is no ice are masked. This is done on
the loop on output streams (with loop index n).

This special averaging is done by accessing a2D and a3Dc using the
variable n_<var>(ns), which corresponds to the index in the
avail_hist_field array where this history variable/frequency is defined.
By construction, this index correponds to the loop index 'n', for both
the 2D and the 3D loops. Simplify the code by using 'n' directly.