ice_history: use avg_ice_present, mask_ice_free_points to reduce duplication
Some CMIP variables are processed differently in
ice_history::accum_hist: they are averaged only for time steps when ice
is present, and points where ice is absent are masked. This processing
is repeated for each of these variables in the 2D and 3Dc loops.

To reduce code duplication, use the new components avg_ice_present and
mask_ice_free_points of ice_hist_field to perform this processing only
for variables that were defined accordingly. The relevant variables
already have those components defined as of the previous commit.

Note that we still need a separate loop for the variable 'sialb' (sea
ice albedo) to mask points below the horizon.