doc: ug_implementation.rst: do not use curly quotes
The namelist excerpt in section 'History' of the Implementation part of
the user guide uses curly quotes (’) instead of regular straight quotes
('). This is probably a remnant of the LaTeX version of the doc. These
quotes can't be used in Fortran and so copy pasting from the doc to the
namelist causes runtime failures. Use straigth quotes instead.