doc: ug_implementation.rst: mention histfreq_base and hist_avg are per-stream
In 35ec167d (Add functionality to change hist_avg for each stream
(#827), 2023-05-17), hist_avg was made into an array, allowing each
stream to individually be set to instantaneous or averaged mode. The
first paragraph of the "History" section of the user guide was updated,
but another paragraph a little below was not.

In 933b148c (Extend restart output controls, provide multiple frequency
options (#850), 2023-08-24), histfreq_base was also made into an array,
but the "History" section of the user guide was not updated.

Adjust the wording of the doc to reflect the fact that both hist_avg and
histfreq_base are per-stream. Also adjust the namelist excerpt to make
histfreq_base an array, and align hist_avg with it.