doc: ug_implementation.rst: refer to 'timemanager' after mentioning histfreq_base
In 34dc6670 (Namelist option for time axis position. (#839),
2023-07-06), the namelist option hist_time_axis was added, and the
"History" section of the user guide updated to mention it.

The added sentence, however, separates the mention of 'histfreq_base'
and the reference to the "Time manager" section, which explains the
different allowed values for that variable. Move the reference up so
both are next to each other.