doc: update histfreq_base and hist_avg descriptions (#898)
* doc: ug_implementation.rst: do not use curly quotes

The namelist excerpt in section 'History' of the Implementation part of
the user guide uses curly quotes (’) instead of regular straight quotes
('). This is probably a remnant of the LaTeX version of the doc. These
quotes can't be used in Fortran and so copy pasting from the doc to the
namelist causes runtime failures. Use straigth quotes instead.

* doc: ug_implementation.rst: align histfreq_n with histfreq

Align frequencies with their respective streams, which makes the example
clearer.

* doc: ug_implementation.rst: avoid "now" and "still"

The documentation talks about the current version of the code, so it is
unnecessary to use words like "now" and "still" to talk about the model
features. Remove them.

* doc: ug_implementation.rst: mention histfreq_base and hist_avg are per-stream

In 35ec167d (Add functionality to change hist_avg for each stream
(#827), 2023-05-17), hist_avg was made into an array, allowing each
stream to individually be set to instantaneous or averaged mode. The
first paragraph of the "History" section of the user guide was updated,
but another paragraph a little below was not.

In 933b148c (Extend restart output controls, provide multiple frequency
options (#850), 2023-08-24), histfreq_base was also made into an array,
but the "History" section of the user guide was not updated.

Adjust the wording of the doc to reflect the fact that both hist_avg and
histfreq_base are per-stream. Also adjust the namelist excerpt to make
histfreq_base an array, and align hist_avg with it.

* doc: ug_implementation.rst: refer to 'timemanager' after mentioning histfreq_base

In 34dc6670 (Namelist option for time axis position. (#839),
2023-07-06), the namelist option hist_time_axis was added, and the
"History" section of the user guide updated to mention it.

The added sentence, however, separates the mention of 'histfreq_base'
and the reference to the "Time manager" section, which explains the
different allowed values for that variable. Move the reference up so
both are next to each other.