Split N/E grid computation out of Tlonlat, create NElonlat subroutine.
See https://github.com/CICE-Consortium/CICE/issues/897

When TLON, TLAT, ANGLET are on the CICE grid, Tlonlat is NOT called.  This
meant N and E grid info was never computed.  This would fail during history
writing with invalid values in N and E grid arrays.  And it would also
cause problem if the C-grid were run with this type of CICE grid.

There are no test grids that have TLON, TLAT, ANGLET on them, so this
error was not found in standard test suites.  This was detected by
users.