Split N/E grid computation out of Tlonlat, create NElonlat subroutine. (#899)
* Split N/E grid computation out of Tlonlat, create NElonlat subroutine.

See https://github.com/CICE-Consortium/CICE/issues/897

When TLON, TLAT, ANGLET are on the CICE grid, Tlonlat is NOT called.  This
meant N and E grid info was never computed.  This would fail during history
writing with invalid values in N and E grid arrays.  And it would also
cause problem if the C-grid were run with this type of CICE grid.

There are no test grids that have TLON, TLAT, ANGLET on them, so this
error was not found in standard test suites.  This was detected by
users.

* Add gx3 grid/kmt files with TLON, TLAT, ANGLET netcdf grid test.

The grid and kmt files were produced from a gx3 history file.  Results
are not bit-for-bit with the standard gx3 runs, but seem to be roundoff
different initially (as expected).