ug_testing.rst: also mention checking the base suite results
In the "End-To-End Testing Procedure" section of the user guide, we
instruct users to run a base suite and a test suite, but only mention
checking the results of the test suite.

Also mention checking the results of the base suite first, to make sure
everything passes before checking the test suite.

Suggested-by: Jean-Francois Lemieux <jean-francois.lemieux@canada.ca>