Typo in nmode0 in PIO for CESM_COUPLED (#938)
For CESMCOUPLED, the pio_iotype is set though CMEPS. For completeness, we also need to check the ioformat set through CMEPS. ioformat is used to set the nmode0. nmode0 and clobber combine to set the nmode flag for pio_createfile / pio_openfile operations. Due a typo nmode0 was not being set for netcdf4 iotypes (before this change).