Update the calculation of uvel and vvel in evp dynamics (#953)
Update the calculation of uvel and vvel in subroutine evp in file ice_dyn_evp.F90. Do an unmasked grid_average_X2Y when averaging from uvelE to uvel and from vvelN to vvel instead of a masked average. The masking is handled separately. This change is bit-for-bit and saves a few flops. Closes #952.

Update the documentation to add the hemispheric sign dependence in the water stress terms in the dynamics equations. Closes #951.