Access alternative congelation ice formulation in icepack (#965)
Add support for 'one-step' congelation ice formation in addition to current default 'two-step' option. Added a new namelist, congel_freeze to control the congelation option. This CICE update requires an update to Icepack to support the new congel_freeze option. The default setting for congel_freeze is set to 'two-step' for backwards compatibility. The new 'one-step' option will be evaluated and may become the default option in the future.

Testing indicates the default setting is bit-for-bit identical with the current main.