Confirm backward compatibility for PIO reads on non-parallel files (#966)
Add ability for netcdf file reading to fall back to regular netcdf if a file is initially opened (for instance with hdf5 reading) and fails to open properly.

For some versions of the PIO library (<2.5.10), attempting a parallel read with filetype netcdf4 (hdf5) of a file which does not support parallel reads fails. The expected behaviour (which is fixed in latest PIO versions) is to fall back to a serial read. The change ensures support for older PIO versions.