Allow kmt variable name of "mask" and different size netcdf 2d reads (#985)
This change allows for a single processor read of netcdf variables that are not of size nx_global by ny_global. This is to make it easier to extent the ice_grid routines.

This changes allows the mask variable in the kmt netcdf file to be called "mask" instead of "kmc". Both options now work.

This change splits reading the mask and grid into seperate subroutines, so the routine can be reused independently later.

* Allow for varname=mask and reading different size nc files

* Update ug_implementation.rst

* fix gadi test submission

* Tidy gadi compiler flags