Add pio_finalize calls for CICE history and restart files (#995)
* Add pio_finalize calls for CICE history and restart files when running with PIO.  This
cleans up some PIO internal data.  The code was aborting when trying to write
lots of files (>600) because we reinitialize PIO for each file.  This is a short term
fix.  A longer term fix is to initialize PIO once and reuse as much as possible.

Add a new test that runs CICE for a year writing hourly files (>8000 files).  Test with
pio2 and netcdf.

* Add error checks and CESMCOUPLED cpp to new ice_pio_finalize method.