Fix outgoing lw for history and coupling. (#999)
This is a two-line change in ice_flux.F90 (scale_fluxes) to check for the condition where aice > 0, but flwout > -puny (effectively zero). This means the cell was ice free at the beginning of the step, but non-zero at the end. It sets the outgoing longwave to the open ocean value based on the freezing temperature. This is bfb, but will change the history output of flwup and flwup_ai. Test results are coming.