Update derecho gnu to add coverage capability with lcov (#1003)
Update derecho gnu to support code coverage testing

    Update to ncarenv/24.12 and gcc/12.4.0
    Add coverage compiler flags
    Update lcov scripting as needed

As part of the CICE port to the new coverage tools, a bug was found in the tools that requires debug flags to be disabled when compiling with coverage flags. This was not the case when coverage was running on cheyenne. See linux-test-project/lcov#385