Setting calendar attribute in output to proleptic_gregorian (#1024)
This sets the calendar attribute correctly in history output. The docs and code comments already note the calendar is actually proleptic gregorian. 

The calendar CICE with leap years is now called proleptic_gregorian rather than greogorian. The gregorian calendar starts in 1582, and a calendar which uses the same pattern of leap years as the modern calendar but exists before 1582 should be refered to as proleptic_gregorian .

Support the old gregorian calendar name for backwards compatibility.  This calendar in the proleptic gregorian and always has been.  The true gregorian calendar (where the calendar changes and leap days start in 1582) is not supported and is called the "standard" calendar now in CF conventions.

Add checks for invalid calendars.

---------

Co-authored-by: anton-seaice <anton-seaice@users.noreply.github.com>